Edited Sections and Fixed Small Details
I cleared out sections 3, 4, and 5 since they can be empty for Increment One. I also wrote in your guys' last names. Don't forget to change the version number not only on the document itself, but also the name of the file. We're on v1.2 currently. Next edit will be v1.3.